// Test.Extension/Client/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type CreateTestItemRequest = {
    name: string;
    description: string;
};

export enum EventMessageTypeModel {
    DEFAULT = 'Default',
    INFO = 'Info',
    ERROR = 'Error',
    SUCCESS = 'Success',
    WARNING = 'Warning'
}

export type NotificationHeaderModel = {
    message: string;
    category: string;
    type: EventMessageTypeModel;
};

export type TestItem = {
    id: number;
    name: string;
    description: string;
    createdAt: string;
    updatedAt?: string | null;
};

export type UpdateTestItemRequest = {
    name: string;
    description: string;
};

export type GetItemsData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/umbraco/mindburn/api/v1/items';
};

export type GetItemsErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
};

export type GetItemsResponses = {
    /**
     * OK
     */
    200: Array<TestItem>;
};

export type GetItemsResponse = GetItemsResponses[keyof GetItemsResponses];

export type CreateItemData = {
    body?: CreateTestItemRequest;
    path?: never;
    query?: never;
    url: '/umbraco/mindburn/api/v1/items';
};

export type CreateItemErrors = {
    /**
     * Bad Request
     */
    400: unknown;
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
};

export type CreateItemResponses = {
    /**
     * Created
     */
    201: TestItem;
};

export type CreateItemResponse = CreateItemResponses[keyof CreateItemResponses];

export type DeleteItemData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/umbraco/mindburn/api/v1/items/{id}';
};

export type DeleteItemErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
    /**
     * Not Found
     */
    404: unknown;
};

export type DeleteItemResponses = {
    /**
     * No Content
     */
    204: void;
};

export type DeleteItemResponse = DeleteItemResponses[keyof DeleteItemResponses];

export type GetItemData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/umbraco/mindburn/api/v1/items/{id}';
};

export type GetItemErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
    /**
     * Not Found
     */
    404: unknown;
};

export type GetItemResponses = {
    /**
     * OK
     */
    200: TestItem;
};

export type GetItemResponse = GetItemResponses[keyof GetItemResponses];

export type UpdateItemData = {
    body?: UpdateTestItemRequest;
    path: {
        id: number;
    };
    query?: never;
    url: '/umbraco/mindburn/api/v1/items/{id}';
};

export type UpdateItemErrors = {
    /**
     * Bad Request
     */
    400: unknown;
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
    /**
     * Not Found
     */
    404: unknown;
};

export type UpdateItemResponses = {
    /**
     * OK
     */
    200: TestItem;
};

export type UpdateItemResponse = UpdateItemResponses[keyof UpdateItemResponses];

export type ClientOptions = {
    baseUrl: 'http://localhost:14327' | (string & {});
};